Move run function and tests to folder
