+++ src/run/run.test.js
@@ -1,14 +1,14 @@
+import waitForDeployCreation from '../wait-for-deploy-creation'
+import waitForReadiness from '../wait-for-readiness'
+import waitForUrl from '../wait-for-url'
 import * as fakes from './fakes'
 import run from './run'
-import waitForDeployCreation from './wait-for-deploy-creation'
-import waitForReadiness from './wait-for-readiness'
-import waitForUrl from './wait-for-url'
 import * as core from '@actions/core'
 import * as github from '@actions/github'
 
-jest.mock('./wait-for-deploy-creation')
-jest.mock('./wait-for-readiness')
-jest.mock('./wait-for-url')
+jest.mock('../wait-for-deploy-creation')
+jest.mock('../wait-for-readiness')
+jest.mock('../wait-for-url')
 jest.mock('@actions/core')
 
 const defaultDeployTimeout = 300
